requirements.txt
pillow
streamlit 
pytesseract